refactor: fix pr comments
